+++ projects/ids-enterprise-ng/src/lib/arrange/index.ts
@@ -1,0 +1,2 @@
+export * from './soho-arrange.directive';
+export * from './soho-arrange.module';